Trait tor_basic_utils::rangebounds::RangeBoundsExt

pub trait RangeBoundsExt<T>: RangeBounds<T> {
    // Required method
    fn intersect<'a, U: RangeBounds<T>>(
        &'a self,
        other: &'a U
    ) -> Option<(Bound<&'a T>, Bound<&'a T>)>;
}

An extension trait for RangeBounds.

Required Methods

fn intersect<'a, U: RangeBounds<T>>( &'a self, other: &'a U ) -> Option<(Bound<&'a T>, Bound<&'a T>)>

Compute the intersection of two RangeBounds.

In essence, this computes the intersection of the intervals described by bounds of the two objects.

Returns None if the intersection of the two ranges is the empty set.